Make DirectoryCache fail-open with stale data and async update
Previously, if DirectoryCache encountered an HTTP issue while updating,
it would sleep for one second and then retry, up to three times, before
raising the error to the consumer.

This has two problems:

  1. It blocks the main thread for ~3 seconds for a persistent failure
  2. The consumer needs to also rescue some of these exceptions and
     decide what to do with it.

This commit alters the default journey. If the DirectoryCache is unable
to refresh itself immediately, we return stale data (or no data, if none
available), log the issue, then enqueue a job to try and fetch the data
in a while. This background job will then continually retry to update
the DirectoryCache until it succeeds.

After this change, the consumer can expect the cache to be immediately
updated or receive stale data, and not have to worry about catching
exceptions or blocking the main thread for very long.