Fix NoMethodError in implementation of DirectoryCache.fetch
The previous code said:

  json = Oj.load(cache) if cache

The problem with this is that `json` is `nil` if the cache fails (e.g.
when Redis is down). However, further down in the method are several
places which are actually expecting the JSON to be a hash.

I've rewritten the method to return early if this is the cache is
unavailable, I hope this can fix the problem.